Rename files, add some instructions to markdown.js
