+++ shared/naturalcrit/codeEditor/codeEditor.less
@@ -5,8 +5,8 @@
 @import (less) 'codemirror/addon/hint/show-hint.css';
 
 //Icon fonts included so they can appear in emoji autosuggest dropdown
-@import (less) './themes/fonts/icon fonts/diceFont.less';
-@import (less) './themes/fonts/icon fonts/elderberryInn.less';
+@import (less) './themes/fonts/iconFonts/diceFont.less';
+@import (less) './themes/fonts/iconFonts/elderberryInn.less';
 
 @keyframes sourceMoveAnimation {
 	50%  {background-color: red;   color: white;}
