+++ themes/V3/5ePHB/style.less
@@ -1,6 +1,6 @@
 @import (less) './themes/assets/assets.less';
-@import (less) './themes/fonts/icon fonts/elderberryInn.less';
-@import (less) './themes/fonts/icon fonts/diceFont.less';
+@import (less) './themes/fonts/iconFonts/elderberryInn.less';
+@import (less) './themes/fonts/iconFonts/diceFont.less';
 
 :root {
 	//Colors
