+++ themes/V3/Blank/style.less
@@ -1,6 +1,6 @@
 @import (less) './themes/fonts/5e/fonts.less';
 @import (less) './themes/assets/assets.less';
-@import (less) './themes/fonts/icon fonts/diceFont.less';
+@import (less) './themes/fonts/iconFonts/diceFont.less';
 
 :root {
 	//Colors
