+++ themes/fonts/iconFonts/diceFont.js
