+++ themes/fonts/iconFonts/diceFont.less
@@ -3,7 +3,7 @@
 	font-family : 'DiceFont';
 	font-style : normal;
 	font-weight : normal;
-	src : url('../../../fonts/icon fonts/diceFont.woff2');
+	src : url('../../../fonts/iconFonts/diceFont.woff2');
 } 
 
 .df {
