+++ themes/fonts/iconFonts/elderberryInn.less
@@ -3,7 +3,7 @@
 	font-family : 'Elderberry-Inn';
 	font-style : normal;
 	font-weight : normal;
-	src : url('../../../fonts/icon fonts/elderberryInn.woff2');
+	src : url('../../../fonts/iconFonts/elderberryInn.woff2');
 } 
 
 .ei {
